requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
torch==2.5.1
transformers==4.49.0
accelerate==1.2.1
nltk==3.9.1
tqdm==4.67.1
matplotlib==3.10.0
PyYAML==6.0.2
safetensors==0.5.3
pytest==8.3.5

